chore: add embed
